+++ packages/@expo/config-plugins/src/index.ts
@@ -5,7 +5,12 @@ import * as AndroidConfig from './android';
 import * as IOSConfig from './ios';
 import { provider, withGeneratedBaseMods } from './plugins/createBaseMod';
 import { getAndroidModFileProviders, withAndroidBaseMods } from './plugins/withAndroidBaseMods';
-import { getIosModFileProviders, withIosBaseMods } from './plugins/withIosBaseMods';
+import {
+  getIosModFileProviders,
+  getMacosModFileProviders,
+  withIosBaseMods,
+  withMacosBaseMods,
+} from './plugins/withIosBaseMods';
 import * as XML from './utils/XML';
 import * as CodeGenerator from './utils/generateCode';
 import * as History from './utils/history';
@@ -71,4 +76,6 @@ export const BaseMods = {
   getAndroidModFileProviders,
   withIosBaseMods,
   getIosModFileProviders,
+  withMacosBaseMods,
+  getMacosModFileProviders,
 };